#282 change module name to "chrono" and moved creg timeline references to examples file
