+++ src/chrono/examples/creg_timeline_config.json
@@ -1,0 +1,162 @@
+{
+    "c400_config": 7,
+    "hours_config": [
+        [
+            "0-12am",
+            60
+        ],
+        [
+            "1-1am",
+            120
+        ],
+        [
+            "2-2am",
+            180
+        ],
+        [
+            "3-3am",
+            240
+        ],
+        [
+            "4-4am",
+            300
+        ],
+        [
+            "5-5am",
+            360
+        ],
+        [
+            "6-6am",
+            420
+        ],
+        [
+            "7-7am",
+            480
+        ],
+        [
+            "8-8am",
+            540
+        ],
+        [
+            "9-9am",
+            600
+        ],
+        [
+            "10-10am",
+            660
+        ],
+        [
+            "11-11am",
+            720
+        ],
+        [
+            "12-12pm",
+            780
+        ],
+        [
+            "13-1pm",
+            840
+        ],
+        [
+            "14-2pm",
+            900
+        ],
+        [
+            "15-3pm",
+            960
+        ],
+        [
+            "16-4pm",
+            1020
+        ],
+        [
+            "17-5pm",
+            1080
+        ],
+        [
+            "18-6pm",
+            1140
+        ],
+        [
+            "19-7pm",
+            1200
+        ],
+        [
+            "20-8pm",
+            1260
+        ],
+        [
+            "21-9pm",
+            1320
+        ],
+        [
+            "22-10pm",
+            1380
+        ],
+        [
+            "23-11pm",
+            1440
+        ]
+    ],
+    "months_config": [
+        [
+            "mar",
+            31
+        ],
+        [
+            "apr",
+            59
+        ],
+        [
+            "may",
+            90
+        ],
+        [
+            "jun",
+            120
+        ],
+        [
+            "jul",
+            151
+        ],
+        [
+            "aug",
+            181
+        ],
+        [
+            "sep",
+            212
+        ],
+        [
+            "oct",
+            243
+        ],
+        [
+            "nov",
+            273
+        ],
+        [
+            "dec",
+            304
+        ],
+        [
+            "jan",
+            334
+        ],
+        [
+            "feb",
+            365
+        ]
+    ],
+    "timeline_label": "creg",
+    "weekdays_config": [
+        "Wednesday",
+        "Thursday",
+        "Friday",
+        "Saturday",
+        "Sunday",
+        "Monday",
+        "Tuesday"
+    ],
+    "yr1_jan1_offset": 440640
+}
