+++ src/chrono/test/test_cregtime_config.py
@@ -1,9 +1,8 @@
-from src._road.road import RoadUnit, is_string_in_road
 from src.bud.group import awardlink_shop
 from src.bud.reason_idea import reasonunit_shop
-from src.bud.idea import IdeaUnit, ideaunit_shop
-from src.bud.bud import budunit_shop, BudUnit
-from src.creg.creg import (
+from src.bud.idea import ideaunit_shop
+from src.bud.bud import budunit_shop
+from src.chrono.examples.chrono_examples import (
     add_time_creg_ideaunit,
     get_creg_min_from_dt,
     get_cregtime_text,
@@ -19,7 +18,7 @@ from src.creg.creg import (
     creg_months_list,
     creg_weekday_ideaunits,
 )
-from src.creg.timebuilder import (
+from src.chrono.timebuilder import (
     day_length,
     week_length,
     time_str,
